refactor: merge ReactiveManager into manager.py
ReactiveManager (90 lines) was a thin subclass of BuilderManager
in its own file. Merge into manager.py for cohesion.
